+++ compages/_structure.py
@@ -33,17 +33,25 @@ def collect_messages(
     return result
 
 
+class SequentialStructureHandler(ABC):
+    @abstractmethod
+    def applies(self, structure_into: Any, val: Any) -> bool: ...
+
+    @abstractmethod
+    def __call__(self, structurer: "Structurer", structure_into: Any, val: Any) -> Any: ...
+
+
 _T = TypeVar("_T")
 
 
 class Structurer:
     def __init__(
         self,
-        handlers: Mapping[Any, Callable[["Structurer", type, Any], Any]] = {},
-        predicate_handlers: Iterable["PredicateStructureHandler"] = [],
+        lookup_handlers: Mapping[Any, Callable[["Structurer", type, Any], Any]] = {},
+        sequential_handlers: Iterable[SequentialStructureHandler] = [],
     ):
-        self._handlers = handlers
-        self._predicate_handlers = predicate_handlers
+        self._lookup_handlers = lookup_handlers
+        self._sequential_handlers = sequential_handlers
 
     @overload
     def structure_into(self, structure_into: NewType, val: Any) -> Any: ...
@@ -56,16 +64,16 @@ class Structurer:
         lookup_order = get_lookup_order(structure_into)
 
         for tp in lookup_order:
-            handler = self._handlers.get(tp, None)
+            handler = self._lookup_handlers.get(tp, None)
             if stack.push(handler):
                 return stack.result()
 
-        # Check all predicate handlers in order and see if there is one that applies
+        # Check all sequential handlers in order and see if there is one that applies
         # TODO (#10): should `applies()` raise an exception which we could collect
         # and attach to the error below, to provide more context on why no handlers were found?
-        for predicate_handler in self._predicate_handlers:
-            if predicate_handler.applies(structure_into, val):
-                if stack.push(predicate_handler):
+        for sequential_handler in self._sequential_handlers:
+            if sequential_handler.applies(structure_into, val):
+                if stack.push(sequential_handler):
                     return stack.result()
                 break
 
@@ -75,11 +83,3 @@ class Structurer:
         raise StructuringError(
             f"Could not find a non-generator handler to structure into {structure_into}"
         )
-
-
-class PredicateStructureHandler(ABC):
-    @abstractmethod
-    def applies(self, structure_into: Any, val: Any) -> bool: ...
-
-    @abstractmethod
-    def __call__(self, structurer: Structurer, structure_into: Any, val: Any) -> Any: ...